# blogRESTAPIdocumentation.yaml
openapi: 3.0.3
info:
  title: Blog REST APIs using Golang and SQLITE
  description: |-
    This is a sample Blog Server based on the OpenAPI 3.0 specification.
  contact:
    email: sujithsureshkumar081@gmail.com
  version: 1.0.11
servers:
  - url: http://localhost:8080/api
...
paths:
...
  /tag/{tagId}:
    get:
      tags:
        - tag
      summary: Find tag by ID
      description: Returns a single Tag
      operationId: getTagById
      parameters:
        - name: tagId
          in: path
          description: ID of tag to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagResponse'          
            application/xml:
              schema:
                $ref: '#/components/schemas/TagResponse'
        '400':
          description: Invalid ID supplied
        '404':
          description: Page not found
    put:
      tags:
        - tag
      summary: Updates a Tag by Id
      description: ''
      operationId: updateTag
      parameters:
        - name: tagId
          in: path
          description: ID of the tag that needs to be updated
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Create a new Tag 
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tag'
          application/xml:
            schema:
              $ref: '#/components/schemas/Tag'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Tag'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'          
            application/xml:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '405':
          description: Invalid input
    delete:
      tags:
        - tag
      summary: Deletes a tag by Id
      description: delete a tag by id
      operationId: deleteTag
      parameters:
        - name: tagId
          in: path
          description: tag id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'          
            application/xml:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Invalid tag value
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
          example: Success
...
    Tag:
      required:
        - tag
      type: object
      properties:
        tag:
          type: string
          example: iphone
        author_id:
          type: string
          example: 8
        post_id:
          type: string
          example: 9
      xml:
        name: Tag
    TagResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            id:
             type: integer
             example: 8
            tag:
             type: string
             example: kerala
            post_id:
             type: string
             example: 6
            author_id:
             type: string
             example: 7
      xml:
        name: Author
...